Extended function TSnap::PrintInfo() in gbase.h

+++ snap-core/util.cpp
@@ -705,7 +705,7 @@ int WriteN(int fd, char *ptr, int nbytes) {
 
   nleft = nbytes;
   while (nleft > 0) {
-    nwritten = write(fd, ptr, nleft);
+    nwritten = (int) write(fd, ptr, nleft);
     if (nwritten <= 0) {
       return nwritten;
     }
